+++ src/policies/exponential_backoff.rs
@@ -1,28 +1,60 @@
-use crate::{RetryDecision, RetryPolicy};
-use chrono::Utc;
+use crate::{Jitter, RetryDecision, RetryPolicy};
+use chrono::{DateTime, Utc};
 use rand::distributions::uniform::{UniformFloat, UniformSampler};
-use std::{cmp, time::Duration};
+use std::{
+    cmp::{self, min},
+    time::Duration,
+};
 
-const MIN_JITTER: f64 = 0.0;
-const MAX_JITTER: f64 = 3.0;
-
-/// We are using the "decorrelated jitter" approach detailed here:
-/// <https://aws.amazon.com/blogs/architecture/exponential-backoff-and-jitter/>
+/// Exponential backoff with optional jitter.
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
+#[non_exhaustive]
 pub struct ExponentialBackoff {
     /// Maximum number of allowed retries attempts.
-    pub max_n_retries: u32,
-    /// Minimum waiting time between two retry attempts (it can end up being lower due to jittering).
+    pub max_n_retries: Option<u32>,
+    /// Minimum waiting time between two retry attempts (it can end up being lower when using full jitter).
     pub min_retry_interval: Duration,
     /// Maximum waiting time between two retry attempts.
     pub max_retry_interval: Duration,
-    /// Growing factor governing how fast the retry interval increases with respect to the number
-    /// of failed attempts. If set to 3:
-    /// - first retry: 3^0 = 1
-    /// - second retry: 3^1 = 3
-    /// - third retry: 3^2 = 9
-    /// ...
-    pub backoff_exponent: u32,
+    /// How we apply jitter to the calculated backoff intervals.
+    pub jitter: Jitter,
+}
+
+/// Exponential backoff with a maximum retry duration.
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub struct ExponentialBackoffTimed {
+    /// Maximum duration the retries can continue for, after which retries will stop.
+    max_total_retry_duration: Duration,
+
+    backoff: ExponentialBackoff,
+}
+
+/// Exponential backoff with a maximum retry duration, for a task with a known start time.
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub struct ExponentialBackoffWithStart {
+    started_at: DateTime<Utc>,
+
+    inner: ExponentialBackoffTimed,
+}
+
+/// Builds an exponential backoff policy.
+///
+/// # Example
+///
+/// ```rust
+/// use retry_policies::{RetryDecision, RetryPolicy, Jitter};
+/// use retry_policies::policies::ExponentialBackoff;
+/// use std::time::Duration;
+///
+/// let backoff = ExponentialBackoff::builder()
+///     .retry_bounds(Duration::from_secs(1), Duration::from_secs(60))
+///     .jitter(Jitter::Bounded)
+///     .build_with_total_retry_duration(Duration::from_secs(24 * 60 * 60));
+/// ```
+pub struct ExponentialBackoffBuilder {
+    min_retry_interval: Duration,
+    max_retry_interval: Duration,
+    jitter: Jitter,
 }
 
 impl ExponentialBackoff {
@@ -34,31 +66,50 @@ impl ExponentialBackoff {
     /// use std::time::Duration;
     ///
     /// let backoff = ExponentialBackoff::builder()
-    ///     .build_with_total_retry_duration(Duration::from_secs(24 * 60 * 60));
+    ///     .build_with_max_retries(5);
     ///
-    /// assert_eq!(backoff.backoff_exponent, 3);
     /// assert_eq!(backoff.min_retry_interval, Duration::from_secs(1));
     /// assert_eq!(backoff.max_retry_interval, Duration::from_secs(30 * 60));
-    /// assert_eq!(backoff.max_n_retries, 55); // calculated
+    /// assert_eq!(backoff.max_n_retries, Some(5));
     /// ```
     pub fn builder() -> ExponentialBackoffBuilder {
         <_>::default()
+    }
+
+    fn too_many_attempts(&self, n_past_retries: u32) -> bool {
+        self.max_n_retries
+            .is_some_and(|max_n| max_n <= n_past_retries)
     }
 }
 
 impl RetryPolicy for ExponentialBackoff {
     fn should_retry(&self, n_past_retries: u32) -> RetryDecision {
-        if n_past_retries >= self.max_n_retries {
+        if self.too_many_attempts(n_past_retries) {
             RetryDecision::DoNotRetry
         } else {
-            let unjittered_wait_for = self.min_retry_interval
-                * self
-                    .backoff_exponent
-                    .checked_pow(n_past_retries)
-                    .unwrap_or(u32::MAX);
-            let jitter_factor =
-                UniformFloat::<f64>::sample_single(MIN_JITTER, MAX_JITTER, &mut rand::thread_rng());
-            let jittered_wait_for = unjittered_wait_for.mul_f64(jitter_factor);
+            let unjittered_wait_for = min(
+                self.max_retry_interval,
+                self.min_retry_interval * 2_u32.checked_pow(n_past_retries).unwrap_or(u32::MAX),
+            );
+
+            let jittered_wait_for = match self.jitter {
+                Jitter::None => unjittered_wait_for,
+                Jitter::Full => {
+                    let jitter_factor =
+                        UniformFloat::<f64>::sample_single(0.0, 1.0, &mut rand::thread_rng());
+
+                    unjittered_wait_for.mul_f64(jitter_factor)
+                }
+                Jitter::Bounded => {
+                    let jitter_factor =
+                        UniformFloat::<f64>::sample_single(0.0, 1.0, &mut rand::thread_rng());
+
+                    let jittered_wait_for =
+                        (unjittered_wait_for - self.min_retry_interval).mul_f64(jitter_factor);
+
+                    jittered_wait_for + self.min_retry_interval
+                }
+            };
 
             let execute_after =
                 Utc::now() + clip_and_convert(jittered_wait_for, self.max_retry_interval);
@@ -67,17 +118,45 @@ impl RetryPolicy for ExponentialBackoff {
     }
 }
 
-/// Clip to the maximum allowed retry interval and convert to chrono::Duration
+/// Clip to the maximum allowed retry interval and convert to [`chrono::Duration`].
 fn clip_and_convert(duration: Duration, max_duration: Duration) -> chrono::Duration {
     // Unwrapping is fine given that we are guaranteed to never exceed the maximum retry interval
-    // in magnitude and that is well withing range for chrono::Duration
+    // in magnitude and that is well within range for chrono::Duration
     chrono::Duration::from_std(cmp::min(duration, max_duration)).unwrap()
 }
 
-pub struct ExponentialBackoffBuilder {
-    min_retry_interval: Duration,
-    max_retry_interval: Duration,
-    backoff_exponent: u32,
+impl ExponentialBackoffTimed {
+    /// Create a [`RetryPolicy`] for a task started at the given
+    pub fn for_task_started_at(&self, started_at: DateTime<Utc>) -> ExponentialBackoffWithStart {
+        ExponentialBackoffWithStart {
+            inner: *self,
+            started_at,
+        }
+    }
+}
+
+impl ExponentialBackoffWithStart {
+    fn trying_for_too_long(&self) -> bool {
+        self.inner.max_total_retry_duration <= Self::elapsed(self.started_at)
+    }
+
+    fn elapsed(started_at: DateTime<Utc>) -> Duration {
+        Utc::now()
+            .signed_duration_since(started_at)
+            .to_std()
+            // If `started_at` is in the future then return a zero duration.
+            .unwrap_or_default()
+    }
+}
+
+impl RetryPolicy for ExponentialBackoffWithStart {
+    fn should_retry(&self, n_past_retries: u32) -> RetryDecision {
+        if self.trying_for_too_long() {
+            RetryDecision::DoNotRetry
+        } else {
+            self.inner.backoff.should_retry(n_past_retries)
+        }
+    }
 }
 
 impl Default for ExponentialBackoffBuilder {
@@ -85,7 +164,7 @@ impl Default for ExponentialBackoffBuilder {
         Self {
             min_retry_interval: Duration::from_secs(1),
             max_retry_interval: Duration::from_secs(30 * 60),
-            backoff_exponent: 3,
+            jitter: Jitter::Full,
         }
     }
 }
@@ -110,77 +189,60 @@ impl ExponentialBackoffBuilder {
         self
     }
 
-    /// Set backoff exponent. _Default 3_.
-    ///
-    /// See [`ExponentialBackoff::backoff_exponent`].
-    pub fn backoff_exponent(mut self, exponent: u32) -> Self {
-        self.backoff_exponent = exponent;
+    /// Set what type of jitter to apply.
+    pub fn jitter(mut self, jitter: Jitter) -> Self {
+        self.jitter = jitter;
         self
     }
 
-    /// Builds a [`ExponentialBackoff`] with the given maximum retries.
+    /// Builds an [`ExponentialBackoff`] with the given maximum retries.
     ///
     /// See [`ExponentialBackoff::max_n_retries`].
     pub fn build_with_max_retries(self, n: u32) -> ExponentialBackoff {
         ExponentialBackoff {
             min_retry_interval: self.min_retry_interval,
             max_retry_interval: self.max_retry_interval,
-            backoff_exponent: self.backoff_exponent,
-            max_n_retries: n,
+            max_n_retries: Some(n),
+            jitter: self.jitter,
         }
     }
 
-    /// Builds a [`ExponentialBackoff`] with [`ExponentialBackoff::max_n_retries`] calculated
-    /// from an approximate total duration. So that after the resultant `max_n_retries` we'll
-    /// have (generally) retried past the given `total_duration`.
+    /// Builds an [`ExponentialBackoff`] with the given maximum total duration for which retries will
+    /// continue to be performed.
     ///
-    /// The _actual_ duration will be approximate due to retry jitter, though this calculation
-    /// is itself deterministic (based off mean jitter).
+    /// Requires the use of [`ExponentialBackoffTimed::for_task_started_at()`].
     ///
     /// # Example
-    /// ```
+    ///
+    /// ```rust
+    /// use chrono::{DateTime, Utc};
+    /// use retry_policies::{RetryDecision, RetryPolicy};
     /// use retry_policies::policies::ExponentialBackoff;
     /// use std::time::Duration;
     ///
     /// let backoff = ExponentialBackoff::builder()
-    ///     .backoff_exponent(2)
-    ///     .retry_bounds(Duration::from_secs(1), Duration::from_secs(60 * 60))
-    ///     // set a retry count so we retry for ~3 hours
-    ///     .build_with_total_retry_duration(Duration::from_secs(3 * 60 * 60));
+    ///     .build_with_total_retry_duration(Duration::from_secs(24 * 60 * 60));
     ///
-    /// // mean delay steps: 1.5s, 3s, 6s, 12s, 24s, 48s, 96s, 192s,
-    /// //                   384s, 768s, 1536s, 3072s, 3600s, 3600s
-    /// // expected total delay: 13342.5s = 3.7h (least number of retries >= 3h)
-    /// assert_eq!(backoff.max_n_retries, 14);
+    /// let started_at = Utc::now()
+    ///     .checked_sub_signed(chrono::Duration::seconds(25 * 60 * 60))
+    ///     .unwrap();
+    ///
+    /// backoff.for_task_started_at(started_at)
+    ///     .should_retry(0); // RetryDecision::DoNotRetry
     /// ```
-    pub fn build_with_total_retry_duration(self, total_duration: Duration) -> ExponentialBackoff {
-        let mut out = self.build_with_max_retries(0);
-
-        const MEAN_JITTER: f64 = (MIN_JITTER + MAX_JITTER) / 2.0;
-
-        let delays = (0u32..).into_iter().map(|n| {
-            let min_interval = out.min_retry_interval;
-            let backoff_factor = out.backoff_exponent.checked_pow(n).unwrap_or(u32::MAX);
-            let n_delay = (min_interval * backoff_factor).mul_f64(MEAN_JITTER);
-            cmp::min(n_delay, out.max_retry_interval)
-        });
-
-        let mut approx_total = Duration::from_secs(0);
-        for (n, delay) in delays.enumerate() {
-            approx_total += delay;
-            if approx_total >= total_duration {
-                out.max_n_retries = (n + 1) as _;
-                break;
-            } else if delay == out.max_retry_interval {
-                // Optimisation: The delay aint changing now
-                let remaining_s = (total_duration - approx_total).as_secs_f64();
-                let additional_tries = (remaining_s / delay.as_secs_f64()).ceil() as usize;
-                out.max_n_retries = (n + 1 + additional_tries) as _;
-                break;
-            }
+    pub fn build_with_total_retry_duration(
+        self,
+        total_duration: Duration,
+    ) -> ExponentialBackoffTimed {
+        ExponentialBackoffTimed {
+            max_total_retry_duration: total_duration,
+            backoff: ExponentialBackoff {
+                min_retry_interval: self.min_retry_interval,
+                max_retry_interval: self.max_retry_interval,
+                max_n_retries: None,
+                jitter: self.jitter,
+            },
         }
-
-        out
     }
 }
 
@@ -191,10 +253,10 @@ mod tests {
 
     fn get_retry_policy() -> ExponentialBackoff {
         ExponentialBackoff {
-            max_n_retries: 6,
+            max_n_retries: Some(6),
             min_retry_interval: Duration::from_secs(1),
             max_retry_interval: Duration::from_secs(5 * 60),
-            backoff_exponent: 3,
+            jitter: Jitter::Full,
         }
     }
 
@@ -202,8 +264,8 @@ mod tests {
     fn if_n_past_retries_is_below_maximum_it_decides_to_retry() {
         // Arrange
         let policy = get_retry_policy();
-        let n_past_retries = (0..policy.max_n_retries).fake();
-        assert!(n_past_retries < policy.max_n_retries);
+        let n_past_retries = (0..policy.max_n_retries.unwrap()).fake();
+        assert!(n_past_retries < policy.max_n_retries.unwrap());
 
         // Act
         let decision = policy.should_retry(n_past_retries);
@@ -216,8 +278,8 @@ mod tests {
     fn if_n_past_retries_is_above_maximum_it_decides_to_mark_as_failed() {
         // Arrange
         let policy = get_retry_policy();
-        let n_past_retries = (policy.max_n_retries..).fake();
-        assert!(n_past_retries >= policy.max_n_retries);
+        let n_past_retries = (policy.max_n_retries.unwrap()..).fake();
+        assert!(n_past_retries >= policy.max_n_retries.unwrap());
 
         // Act
         let decision = policy.should_retry(n_past_retries);
@@ -233,7 +295,7 @@ mod tests {
         let max_interval = chrono::Duration::from_std(policy.max_retry_interval).unwrap();
 
         // Act
-        let decision = policy.should_retry(policy.max_n_retries - 1);
+        let decision = policy.should_retry(policy.max_n_retries.unwrap() - 1);
 
         // Assert
         match decision {
@@ -247,8 +309,7 @@ mod tests {
     #[test]
     fn overflow_backoff_exponent_does_not_cause_a_panic() {
         let policy = ExponentialBackoff {
-            max_n_retries: u32::MAX,
-            backoff_exponent: 2,
+            max_n_retries: Some(u32::MAX),
             ..get_retry_policy()
         };
         let max_interval = chrono::Duration::from_std(policy.max_retry_interval).unwrap();
@@ -271,5 +332,36 @@ mod tests {
     fn builder_invalid_retry_bounds() {
         // bounds are the wrong way round or invalid
         ExponentialBackoff::builder().retry_bounds(Duration::from_secs(3), Duration::from_secs(2));
+    }
+
+    #[test]
+    fn does_not_retry_after_total_retry_duration() {
+        let backoff = ExponentialBackoff::builder()
+            .build_with_total_retry_duration(Duration::from_secs(24 * 60 * 60));
+
+        {
+            let started_at = Utc::now()
+                .checked_sub_signed(chrono::Duration::seconds(23 * 60 * 60))
+                .unwrap();
+
+            let decision = backoff.for_task_started_at(started_at).should_retry(0);
+
+            match decision {
+                RetryDecision::Retry { .. } => {}
+                _ => panic!("should retry"),
+            }
+        }
+        {
+            let started_at = Utc::now()
+                .checked_sub_signed(chrono::Duration::seconds(25 * 60 * 60))
+                .unwrap();
+
+            let decision = backoff.for_task_started_at(started_at).should_retry(0);
+
+            match decision {
+                RetryDecision::DoNotRetry => {}
+                _ => panic!("should not retry"),
+            }
+        }
     }
 }