+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
@@ -6,14 +7,34 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0] - 2023-??-??
+
+### Changed
+
+- [Breaking] Change backoff and jitter algorithms
+  - Change the backoff algorithm to a more conventional exponential backoff.
+  - Replace the decorrelated jitter algorithm with an option of either none, full, or bounded. Defaults to full jitter.
+- [Breaking] Remove `ExponentialBackoffBuilder::backoff_exponent()`
+  - The number of attempts is now used as the exponent.
+- [Breaking] Require a task start time when using a total retry duration
+  - `ExponentialBackoffBuilder::build_with_total_retry_duration()` now returns `ExponentialBackoffTimed` which does not implement `RetryPolicy`.
+- [Breaking] Mark `ExponentialBackoff` as `non_exhaustive`
+  - Can no longer be constructed directly.
+
 ## [0.1.2] - 2022-10-28
+
 ### Added
+
 - `Debug` derived for `RetryDecision`
 
 ## [0.1.1] - 2021-10-18
+
 ### Security
+
 - remove time v0.1 dependency
 
 ## [0.1.0] - 2021-08-11
+
 ### Added
+
 - `ExponentialBackoff` policy.